refactor(adapters): consolidate ADK LLM instrumentation into one module
Wave B2 of the modularisation plan. Pulls the request-side
``LlmRequest`` mutation + measurement surface — which until now lived in
three files — into a single audit-friendly module
``goldfive/adapters/adk_llm_instrumentation.py``.

Moved INTO the new module:
* From ``_adk_dynainst.py`` (deleted): every symbol — the dynamic
  instruction installer + tree walker, the correction-block formatter,
  the readonly-context / session resolution helpers, the LlmAgent
  duck-type, ``install_dynamic_instructions`` /
  ``log_dynamic_instruction_opt_out`` / ``is_dynamic_instruction``,
  ``pending_correction_key``, ``format_correction_block``.
* From ``_adk_plugin.py`` (kept as response-side / callback module):
  ``_measure_request_chars`` (#172 instrumentation),
  ``_RUNTIME_TOOLS_HINT_PREFIX`` / ``_RUNTIME_TOOLS_HINT_END``,
  ``_build_runtime_tools_hint`` / ``_strip_prior_runtime_tools_hint``
  (#168 R3 hint composition),
  ``_apply_agent_max_output_tokens_cap`` +
  ``DEFAULT_AGENT_MAX_OUTPUT_TOKENS`` (#256 ratchet).

Kept SEPARATE as a sibling:
* ``_adk_state_protocol.py`` — the state-audit's
  ``_install_protocol_patch`` patches its ``_set`` and catalogues its
  file path in ``_KNOWN_CALLERS``. Merging it would force a
  catalog-and-patch surface migration that is high-risk and bloats the
  new module unhelpfully. The state-protocol layer is delicate (V2/V3/V4
  migration history) and the audit's stack-walk caller-allowlist encodes
  the file path explicitly.

Backwards compatibility: ``_adk_plugin.py`` re-imports the moved
request-side helpers at module load so in-module references and the
external ``from goldfive.adapters._adk_plugin import …`` callsites
(PromptShaper + several test files) keep resolving without change. The
``_adk_dynainst`` module is deleted; the two production callsites
(``convenience.py`` + ``prompt_shaper.py``) and three test files have
been migrated to import from ``adk_llm_instrumentation``.

Contracts preserved verbatim (no behaviour change):
* observation_only gate semantics — the 4 inject sites still funnel
  through ``PromptShaper`` (Wave B1 #401 wiring untouched).
* state-protocol contract — ``_set`` shim still routes through the
  audit; all goldfive.* key constants stable.
* ``max_output_tokens`` ratchet (#256) — smaller-wins, ceiling <= 0
  opt-out, identical signature.
* ``_strip_prior_runtime_tools_hint`` semantics — strip-then-let-
  PromptShaper-re-inject sequenced contract preserved.

Test plan:
* ``pytest tests/`` → 2530 passed, 59 skipped, 29.70s.
* State-ownership audit under strict mode (autouse fixture
  ``GOLDFIVE_STRICT_STATE_OWNERSHIP=1``) — no ``StateOwnershipViolation``.
* Focused suites for moved surface (test_dynamic_instruction,
  test_runtime_tool_surface_hint, test_correction_injection,
  test_e2e_plan_causal_correction, test_agent_budget, test_state_audit,
  test_adk_adapter, test_prompt_shaper) all pass.

LOC delta:
* ``_adk_dynainst.py``: -468 (deleted)
* ``_adk_plugin.py``: -216 net (helpers extracted, import block added)
* ``_adk_state_protocol.py``: ±0 (one doc cross-reference updated)
* ``adk_llm_instrumentation.py``: +809 (new module)

DO NOT MERGE — team lead reviews.
